Apply applicationError and modify the contract to throw an exception

+++ yggdrash-node/src/test/java/io/yggdrash/node/NodeContractDemoClient.java
@@ -6,7 +6,6 @@ import com.google.gson.JsonObject;
 import io.yggdrash.BlockChainTestUtils;
 import io.yggdrash.ContractTestUtils;
 import io.yggdrash.TestConstants;
-import io.yggdrash.common.config.Constants;
 import io.yggdrash.common.config.DefaultConfig;
 import io.yggdrash.common.contract.ContractVersion;
 import io.yggdrash.common.crypto.HexUtil;
@@ -194,7 +193,7 @@ public class NodeContractDemoClient {
         int times = getSendTimes();
         int amount = 1;
         for (int i = 0; i < times; i++) {
-            JsonObject txBody = ContractTestUtils.transferTxBodyJson("", amount);
+            JsonObject txBody = ContractTestUtils.transferTxBodyJson("", BigInteger.valueOf(amount));
             Transaction tx = createTx(BranchId.of(branchId), txBody);
             sendTransaction(tx);
         }
@@ -348,7 +347,7 @@ public class NodeContractDemoClient {
     private static void sendYeedTx(String address, long amount) {
         int times = getSendTimes();
         for (int i = 0; i < times; i++) {
-            JsonObject txBody = ContractTestUtils.transferTxBodyJson(address, amount);
+            JsonObject txBody = ContractTestUtils.transferTxBodyJson(address, BigInteger.valueOf(amount));
             Transaction tx = createTx(yggdrash, txBody);
             sendTransaction(tx);
         }
